Refactor Vault Sync Logic into VaultSyncManager
This commit introduces `VaultSyncManager` to encapsulate the vault synchronization logic previously handled within `VaultRepositoryImpl`.

Key changes:
- Created `VaultSyncManager` interface and `VaultSyncManagerImpl` implementation.
- Moved sync-related logic, including security stamp checks, organization crypto initialization, and data storage, from `VaultRepositoryImpl` to `VaultSyncManagerImpl`.
- `VaultRepositoryImpl` now delegates sync operations to `VaultSyncManager`.
- Introduced `SecurityStampMismatchException` for clearer error handling.
- Moved `SyncVaultDataResult` to `com.x8bit.bitwarden.data.vault.manager.model`.
- Updated DI modules and tests to reflect these changes.
